requirements.txt
Bottleneck==1.4.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mysql-connector-python==8.4.0
numexpr==2.10.1
numpy==2.0.1
pandas==2.2.3
pip==24.2
protobuf==4.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
setuptools==75.1.0
six==1.16.0
tzdata==2023.3
wheel==0.44.0
